refactor directives into their own files, move index.ts logic to bootstrap.ts and index.ts becomes a clean entry point where you can specify which directives you want to use
